move debug command into its own folder
