+++ projects/gloo/cli/pkg/cmd/debug/debug.go
@@ -1,4 +1,4 @@
-package check
+package debug
 
 import (
 	"bufio"
