+++ projects/gloo/cli/pkg/cmd/debug/debug_suite_test.go
@@ -1,4 +1,4 @@
-package check_test
+package debug_test
 
 import (
 	"testing"
